### 2025-04-24 #8105 @editorialbot
[PRE REVIEW]: VectoRose: A new package for analysing and visualising 3D non-unit vectors in Python
**Submitting author:** @bzrudski (<a href="http://orcid.org/0009-0000-3423-0662">Benjamin Zvi Rudski</a>)
**Repository:** https://github.com/bzrudski/vectorose
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.2.4
