### 2025-04-29 70b09dc
docs: add JOSS badge.
