### 2025-04-30 815f8a5
docs: embed x-frame buffer function `pv.start_xvfb` in try-catch.
To provide a more general solution, instead of checking the platform, we're now doing the slightly more brute-force approach of trying to import and then ignoring if it doesn't work. This helps address differing outputs on macOS.
