### 2025-05-02 0e27257
fix: correct bug in `PolarDiscretiser` that converted spatial coordinates to spherical coordinates instead of vector components.
