### 2025-05-02 55d6dac
0.2.5
Automatically generated by python-semantic-release
