+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "vectorose"
-version = "0.2.4"
+version = "0.2.5"
 description = "Plot polar and spherical histograms from orientation data."
 authors = [
     {name = "Benjamin Z. Rudski", email = "benjamin.rudski@mail.mcgill.ca"},
