### 2025-05-26 2d5ece8
1.0.0
Automatically generated by python-semantic-release
