### 2025-05-26 cfca7ca
build: remove tokens for PyPI and TestPyPI.
