### 2025-05-26 b1e2318
build(ci-cd): update version of `python-semantic-release` action.
