+++ .github/workflows/ci-cd.yml
@@ -81,7 +81,7 @@ jobs:
 
       - name: Use Python Semantic Release to prepare release
         id: release
-        uses: python-semantic-release/python-semantic-release@v8.3.0
+        uses: python-semantic-release/python-semantic-release@v10.0.2
         with:
           github_token: ${{ secrets.GITHUB_TOKEN }}
 
