### 2025-06-04 #8105 @banesullivan
Hey @nkrusch, thanks for considering me to review! I'm not sure if I'll be able to make time for this immediately but I've reached out to some other PyVista developers to see if they might be interested. I'll check back soon if someone is interested or if I can make the time
