### 2025-06-17 #8369 @tkoyama010
> - [ ] **Reproducibility:** If the paper contains original results, results are entirely reproducible by reviewers. If the paper contains no original results, please check this item.

@bzrudski Could you let me know if there is any code in the repository to reproduce Figure 2?
