### 2025-06-23 bzrudski/vectorose#7 @stevenpawley
Consider including datasets used in the tutorials/documentation as package data
Because the datasets used in the documentation for vectorose are small, it could be beneficial to include them as built-in datasets that are accessible by the user, i.e., including them inside of the `vectorose/src` folder rather than externally in the `docs` folder. This would help users to follow and reproduce the examples in the user guide, for example, if the `quickstart_vectors.npy` dataset could be imported using mechanism similar to `sklearn.datasets.load_iris()`. Alternatively, the documentation could emphasize the `vectorose.mock_data` module.

openjournals/joss-reviews#8369
