### 2025-06-24 3def90e
test: convert polar tests to use a fixture to generate the random vectors.
