+++ src/vectorose/polar_data.py
@@ -256,9 +256,7 @@ class PolarDiscretiser:
             bins = self._theta_bins
 
         # Count the vectors in each bin
-        counts = labelled_vectors.groupby(bin_assignment_column).apply(
-            len, include_groups=False
-        )
+        counts = labelled_vectors.groupby(bin_assignment_column).apply(len)
         counts.name = "count"
 
         # Normalise to get the frequencies
