### 2025-06-24 68ea1fe
feat: add sample data in importable module.
