### 2025-06-24 e3e6652
test: add tests for loading sample data.
