### 2025-06-24 #8369 @tkoyama010
LGTM
