### 2025-06-25 93a1516
docs: integrate sample data interface into the documentation examples.
Rename sample datasets to mirror the names used in `vectorose.data` and switch to using the new mechanism for loading sample data.
