### 2025-06-25 ff1ec20
docs: improve documentation for the sample data loader.
