+++ src/vectorose/data/__init__.py
@@ -1,8 +1,13 @@
-"""Sample data
+"""Sample datasets for running VectoRose.
 
 This module grants access to the sample data used in the documentation.
 These datasets may be loaded and accessed as NumPy arrays without requiring
 any external download.
+
+Warnings
+--------
+This module does not exist as an automatic top-level import in vectorose,
+and must therefore be explicitly imported.
 """
 
 import enum
@@ -12,10 +17,16 @@ from .. import io
 import numpy as np
 
 class SampleData(enum.Enum):
-    """Sample dataset.
+    """Interface for accessing sample datasets.
 
     This enumeration provides a list of sample datasets provided with
-    VectoRose, as well as a simple interface to load them.
+    VectoRose, as well as a simple interface to load them. The string
+    values associated with each enumerated instance correspond to the base
+    filename for the respective dataset.
+
+    Warnings
+    --------
+    All datasets are assumed to be stored as a NumPy array (``*.npy``).
 
     Notes
     -----
@@ -30,10 +41,24 @@ class SampleData(enum.Enum):
     """Two clusters with different magnitudes and orientations."""
 
     TWISTED_BLOCKS = "twisted_blocks"
-    """Anisotropy of offset rotated layers of cylinders."""
+    """Anisotropy of offset rotated layers of cylinders.
+    
+    Notes
+    -----
+    Computed using Dragonfly 3D World.
+    """
 
     def load(self) -> np.ndarray:
-        """Load the current dataset to use with VectoRose."""
+        """Load the current dataset to use with VectoRose.
+
+        Returns
+        -------
+        numpy.ndarray
+            Array of shape ``(n, d)`` containing the loaded vectors. If the
+            specific dataset has location coordinates, then ``d == 6``.
+            Otherwise, ``d == 3``. ``n`` represents the number of vectors
+            in the dataset.
+        """
 
         parent_dir = os.path.dirname(__file__)
 
