### 2025-06-25 bzrudski/vectorose#7 @bzrudski
I have bundled the test data and added a simple interface in order to access these datasets. I've also updated the documentation so that the examples mostly use the bundled sample data instead of requiring the user to download the files. @stevenpawley is #8 what you had in mind?
