### 2025-06-25 ebfd52f
docs: remove unnecessary reference from the bibliography file.
