+++ paper/paper.bib
@@ -21,36 +21,6 @@
   urldate = {2024-10-02}
 }
 
-@software{banesullivanPyvistaPyvistaV04232023,
-  title = {Pyvista/Pyvista: V0.42.3},
-  shorttitle = {Pyvista/Pyvista},
-  author = {{Bane Sullivan} and Kaszynski, Alex and Koyama, Tetsuo and {Andras Deak} and {MatthewFlamm} and Favelier, Guillaume and {Jevin Jones} and Larson, Eric and Chiu, Phil and Mologni, Rodrigo and {Dcbr} and {Darikg} and {Beroda} and {RichardScottOZ} and {User27182} and Fernandez, Alex and \AA hl, Henrik and Frei, Simon and Keough, Shannon and Taves, Mike and {Math-Artist} and {Puri} and {ENIAC-6} and {Keurfon Luu} and Neeteson, Nathan and {Arnav} and {Blue Tyson}},
-  date = {2023-10-07},
-  doi = {10.5281/ZENODO.8415866},
-  url = {https://zenodo.org/record/8415866},
-  urldate = {2024-10-02},
-  abstract = {\&lt;!-- Release notes generated using configuration in .github/release.yml at v0.42.3 --\&gt; Patch release with the following fixes/quality of life changes. This being a patch fix, there are no API changes. What's Changed Fix module vtkImageDilateErode3D not found error by @Arnav-2004 \#5020 Set Google.Spacing NO in .vale.ini by @tkoyama010 \#5027 Update geovista integration test by @bjlittle \#4957 Fix GeoVista's CI error due to image regerssion test by @tkoyama010 \#4970 {$<$}strong{$>$}Full Changelog{$<$}/strong{$>$}: https://github.com/pyvista/pyvista/compare/v0.42.2...v0.42.3},
-  organization = {Zenodo},
-  version = {v0.42.3}
-}
-
-@article{bar-sinaiMechanicalStressInduces2016,
-  title = {Mechanical {{Stress Induces Remodeling}} of {{Vascular Networks}} in {{Growing Leaves}}},
-  author = {Bar-Sinai, Yohai and Julien, Jean-Daniel and Sharon, Eran and Armon, Shahaf and Nakayama, Naomi and Adda-Bedia, Mokhtar and Boudaoud, Arezki},
-  editor = {Prusinkiewicz, Przemyslaw},
-  date = {2016-04-13},
-  journaltitle = {PLOS Computational Biology},
-  shortjournal = {PLoS Comput Biol},
-  volume = {12},
-  number = {4},
-  pages = {e1004819},
-  issn = {1553-7358},
-  doi = {10.1371/journal.pcbi.1004819},
-  url = {https://dx.plos.org/10.1371/journal.pcbi.1004819},
-  urldate = {2025-01-22},
-  langid = {english}
-}
-
 @article{beckersUniversalProjectionComputing2011,
   title = {The Universal Projection for Computing Data Carried on the Hemisphere},
   author = {Beckers, Benoit and Masset, Luc and Beckers, Pierre},
@@ -65,14 +35,6 @@
   url = {https://linkinghub.elsevier.com/retrieve/pii/S0010448510002174},
   urldate = {2024-08-16},
   langid = {english}
-}
-
-@software{dawson-haggertyetal.Trimesh2019,
-  title = {Trimesh},
-  author = {{Dawson-Haggerty et al.}},
-  date = {2019-12-08},
-  url = {https://trimesh.org/},
-  version = {3.2.0}
 }
 
 @article{dileepCardiomyocyteOrientationRecovery2023,
@@ -112,41 +74,6 @@
   version = {1.0.6}
 }
 
-@article{hansonVectorCorrelationReview1992,
[…]
