+++ paper/paper.md
@@ -33,7 +33,7 @@ bibliography: paper.bib
 
 # Summary
 
-In scientific analyses, 3D vector orientations are often as important as their magnitudes. For example, in meteorology, wind currents do not simply have a *speed*, but also a *direction* [@klinkComplementaryUseScalar1998]. Meanwhile, in 3D biological structures and assemblies, the (co-)orientation of structural units is a feature of interest that directly reflects structure-function relationships. Previous works have examined the orientation of myocardial fibres in the heart [@lombaertHumanAtlasCardiac2012b; @dileepCardiomyocyteOrientationRecovery2023], while others have focused on the coalignment, or *anisotropy*, of trabecular struts in cancellous bone [@reznikovTechnicalNoteMapping2022]. Whether analysing the vector orientations in isolation or in combination with their magnitudes, numerical representations alone are insufficient to provide a complete understanding of these datasets.
+In scientific analyses, 3D vector orientations are often as important as their magnitudes. For example, in meteorology, wind currents do not simply have a *speed*, but also a *direction* [@klinkComplementaryUseScalar1998]. Meanwhile, in 3D biological structures and assemblies, the (co-)orientation of structural units is a feature of interest that directly reflects structure-function relationships. Previous works have examined the orientation of myocardial fibres in the heart [@lombaertHumanAtlasCardiac2012; @dileepCardiomyocyteOrientationRecovery2023], while others have focused on the coalignment, or *anisotropy*, of trabecular struts in cancellous bone [@reznikovTechnicalNoteMapping2022]. Whether analysing the vector orientations in isolation or in combination with their magnitudes, numerical representations alone are insufficient to provide a complete understanding of these datasets.
 
 To gain such insight, we seek to visualise and quantitatively analyse sets of 3D non-unit vectors. Any intuitive visualisation should represent the distributions of vector magnitude and orientation, as well as the interplay between these two variables, thus enabling observers to identify patterns of dominant orientations and their associated magnitudes. Meanwhile, quantitative insights into these data require *directional statistics* to be applied instead of conventional Euclidean approaches [@fisherStatisticalAnalysisSpherical1993; @mardiaDirectionalStatistics2000]. To meet these needs, in this work, we present **VectoRose**, a new open-source Python package. To facilitate data visualisation of collections of non-unit vectors, VectoRose computes linear vector magnitude histograms, 3D spherical orientation histograms, and a novel type of nested 3D spherical histograms -- simultaneously capturing orientation and magnitude data. Our package also implements existing directional statistics approaches to enable statistical analysis of such vectorial data. VectoRose is agnostic to the workflow used to compute the datasets and can be applied in various scientific fields.
 
