### 2025-06-25 795ab57
build: fix versions for `python-semantic-release`.
