### 2025-06-25 05cfe8d
build: update `python-semantic-release` in `pyproject.toml`.
