+++ pyproject.toml
@@ -36,7 +36,7 @@ pytest-cov = "^4.1.0"
 jupyter = "^1.0.0"
 myst-nb = "^1.0.0"
 sphinx-autoapi = "^3.0.0"
-python-semantic-release = "^8.7.0"
+python-semantic-release = ">=10.1.0"
 ipympl = "^0.9.4"
 pydata-sphinx-theme = "^0.15.3"
 jupyterlab = ">=3"
