### 2025-06-25 6d0658c
build: roll back python-semantic-release versions.
