+++ pyproject.toml
@@ -36,7 +36,6 @@ pytest-cov = "^4.1.0"
 jupyter = "^1.0.0"
 myst-nb = "^1.0.0"
 sphinx-autoapi = "^3.0.0"
-python-semantic-release = ">=10.1.0"
 ipympl = "^0.9.4"
 pydata-sphinx-theme = "^0.15.3"
 jupyterlab = ">=3"
@@ -46,6 +45,7 @@ jupytext = "^1.16.4"
 sphinxcontrib-video = "^0.3.1"
 sphinx-design = "^0.6.1"
 sphinx-copybutton = "^0.5.2"
+python-semantic-release = "^9.21.1"
 
 [tool.semantic_release]
 version_toml = [
