### 2025-06-25 3ecb310
build: roll back python-semantic-release action version.
