### 2025-06-25 af17964
1.0.1
Automatically generated by python-semantic-release
