### 2025-06-25 81817a7
build(ci-cd): revert to token-based publishing on PyPI and Test PyPI.
