### 2025-06-25 6c5d886
feat: add ability to alter the radial axis bounds of polar plot.
