### 2025-06-25 da181bd
1.1.0
Automatically generated by python-semantic-release
