### 2025-06-26 d65c9d8
build: explicitly set the changelog settings.
