### 2025-06-27 bzrudski/vectorose#7 @stevenpawley
Thanks @bzrudski, this looks great!
