### 2025-06-27 bzrudski/vectorose#7 @bzrudski
Thanks! Merged #8 and closing the issue.
