### 2025-06-27 68dccaf
1.2.0
Automatically generated by python-semantic-release
