### 2025-06-27 #8369 @stevenpawley
Hi @bzrudski, in the 'Statement of Need' section, would it be worth adding some comparisons to circular statistics packages outside of the Python community, for example, in the R ecosystem? Those packages tend to contain many statistical tests, but not the visualization capabilities of your package (and I think many are for unit vectors only).
