### 2025-06-27 cc3b224
feat: add ability to colour spherical plots using a log scale.
