### 2025-06-27 079d219
test: add unit test for log scale plotting on spherical histograms.
