### 2025-06-27 8f8118a
fix: loosen some restrictions for log scale plotting.
