### 2025-06-27 1bd1633
fix: set dynamic minimum value for log scale plotting if no minimum is set.
