### 2025-06-27 c166015
docs: add log scale example to the documentation.
