### 2025-06-27 cf32a02
build: per Codecov documentation, add code to upload test results.
