+++ .github/workflows/ci-cd.yml
@@ -36,13 +36,19 @@ jobs:
         uses: pyvista/setup-headless-display-action@v3.2
 
       - name: Run tests
-        run: poetry run pytest tests/ --cov=vectorose --cov-report=xml
+        run: poetry run pytest tests/ --cov=vectorose --cov-report=xml --junitxml=junit.xml -o junit_family=legacy
 
       - name: Use Codecov to track coverage
         uses: codecov/codecov-action@v5.4.2
         with:
             files: ./coverage.xml # Coverage report
             token: ${{ secrets.CODECOV_TOKEN }}
+
+      - name: Upload test results to Codecov
+        if: ${{ !cancelled() }}
+        uses: codecov/test-results-action@v1
+        with:
+          token: ${{ secrets.CODECOV_TOKEN }}
 
       - name: Build documentation
         run: poetry run make html --directory docs/
