### 2025-06-29 7274faa
fix: fix bug that didn't correctly show the scalars with NaNs.
