### 2025-06-29 3de4453
fix: add NaN annotation to the scalar bar.
