### 2025-06-29 5413ff3
1.3.0
Automatically generated by python-semantic-release
