### 2025-07-02 947431d
style: minor code reformatting using `black`.
