### 2025-07-02 7b04b2a
feat: add support for picking mesh cells in `SpherePlotter`.
