+++ src/vectorose/plotting.py
@@ -12,7 +12,7 @@ present in :mod:`.tregenza_sphere`, :mod:`.triangle_sphere` and
 
 import enum
 import functools
-from typing import Any, Iterable, List, Optional
+from typing import Any, Dict, Iterable, List, Optional
 
 import imageio_ffmpeg
 import matplotlib.animation
@@ -28,6 +28,7 @@ import numpy as np
 import pandas as pd
 import vtk
 import pyvista as pv
+from pyvista.plotting.opts import ElementType
 from scipy.spatial.transform import Rotation
 
 from .triangle_sphere import TriangleSphere
@@ -228,6 +229,15 @@ class SpherePlotter:
     _point_label_actor: Optional[vtk.vtkActor2D]
     """The actor controlling the point labels."""
 
+    _picked_cells: List[pv.UnstructuredGrid]
+    """List of picked cells."""
+
+    _picked_cell_actors: Dict[int, pv.Actor]
+    """Actors for picked faces."""
+
+    _picking_active: bool
+    """Indicate whether cell picking is active."""
+
     cmap: str
     """The colour map to use when visualising the data."""
 
@@ -370,6 +380,56 @@ class SpherePlotter:
         return theta
 
     @property
+    def face_picking_active(self) -> bool:
+        """Indicate whether interactive cell picking is active."""
+        return self._picking_active
+
+    @face_picking_active.setter
+    def face_picking_active(self, active: bool):
+        if active:
+            self._plotter.enable_element_picking(
+                self._pick_face,
+                mode=ElementType.CELL,
+                show=False,
+                show_message=False,
+                left_clicking=True,
+            )
+        else:
+            self.clear_picked_cells()
+            self._plotter.disable_picking()
+
+        self._picking_active = active
+
+    @property
+    def picked_cells(self) -> pd.DataFrame:
+        """Get picked face scalar data.
+
+        Returns
+        -------
+        pandas.DataFrame
+            All cell scalar data for the picked mesh cells. These values
+            can then be used by the :class:`.SphereBase` sphere to extract
+            vectors from the picked cells. If no cells are picked, an
+            empty :class:`pandas.DataFrame` is produced.
+        """
+        scalar_data_table = pd.DataFrame()
+
+        for face_index in self._picked_cell_actors.keys():
+            cell = self._picked_cells[face_index]
+
+            scalar_data = cell.cell_data
+
+            column_names = scalar_data.keys()
+            row_values = np.stack(scalar_data.values(), axis=-1)
+
+            cell_df = pd.DataFrame(columns=column_names, data=row_values)
+            scalar_data_table = pd.concat(
+                [scalar_data_table, cell_df], ignore_index=True
+            )
+
+        return scalar_data_table
+
+    @property
     def has_movie_open(self) -> bool:
         """Indicate whether a movie is currently being manually written."""
         return self._has_movie_open
@@ -410,6 +470,11 @@ class SpherePlotter:
 
         # We don't have a movie open when creating the plotter
         self._has_movie_open = False
+
+        # Configure the face selection
+        self._picking_active = False
+        self._picked_cells = []
+        self._picked_cell_actors = {}
 
     def _get_spherical_coordinates_from_camera(self) -> np.ndarray:
         """Get the spherical coordinates for the camera location.
@@ -474,6 +539,32 @@ class SpherePlotter:
         if self._point_label_actor is not None:
             self._point_label_actor.SetVisibility(show_axes)
 
+    def _pick_face(self, cell: pv.UnstructuredGrid):
+        """Callback when a mesh face is picked."""
+
+        # Remove cell if in picked cells
+        if not cell in self._picked_cells:
+            self._picked_cells.append(cell)
[…]
