### 2025-07-02 285a90e
fix: rename the scalar fields for the `TregenzaSphere` and `TriangleSphere`.
