### 2025-07-02 1c54f59
feat: add the ability to filter the vectors from selected histogram cells.
