### 2025-07-02 464fd37
fix: switch clicking to right mouse button for simplicity.
