### 2025-07-02 a6ef9ab
fix: change picked cells to have a consistent magenta colour.
