### 2025-07-02 eded369
refactor: rename functions using cell instead of face.
