### 2025-07-02 fa86e67
test: add tests for cell picking in the `SpherePlotter`.
