+++ tests/test_plotting.py
@@ -7,6 +7,7 @@ This test module is structured to first test the methods in the
 :class:`.SpherePlotter` class, and then the assorted functions for plotting
 using Matplotlib.
 """
+
 import os
 
 import matplotlib as mpl
@@ -104,7 +105,7 @@ def label_vectors_polar(vectors):
 @pytest.fixture
 def mock_nested_histogram_meshes(
     setup_pyvista_environment,
-    label_vectors: tuple[pd.DataFrame, np.ndarray, vr.tregenza_sphere.TregenzaSphere]
+    label_vectors: tuple[pd.DataFrame, np.ndarray, vr.tregenza_sphere.TregenzaSphere],
 ) -> list[pv.PolyData]:
     """Generate several nested dummy meshes for testing."""
 
@@ -114,10 +115,11 @@ def mock_nested_histogram_meshes(
 
     return sphere_meshes
 
+
 @pytest.fixture
 def mock_nested_histogram_meshes_counts(
     setup_pyvista_environment,
-    label_vectors: tuple[pd.DataFrame, np.ndarray, vr.tregenza_sphere.TregenzaSphere]
+    label_vectors: tuple[pd.DataFrame, np.ndarray, vr.tregenza_sphere.TregenzaSphere],
 ) -> list[pv.PolyData]:
     """Generate several nested dummy meshes for testing."""
 
@@ -126,6 +128,22 @@ def mock_nested_histogram_meshes_counts(
     sphere_meshes = sphere.create_histogram_meshes(hist, magnitude_bins)
 
     return sphere_meshes
+
+
+@pytest.fixture
+def mock_orientation_histogram_mesh_counts(
+    setup_pyvista_environment,
+    label_vectors: tuple[pd.DataFrame, np.ndarray, vr.tregenza_sphere.TregenzaSphere],
+) -> pv.PolyData:
+    """Generate several nested dummy meshes for testing."""
+
+    labelled_vectors, magnitude_bins, sphere = label_vectors
+    hist = sphere.construct_marginal_orientation_histogram(
+        labelled_vectors, return_fraction=False
+    )
+    sphere_mesh = sphere.create_shell_mesh(hist)
+
+    return sphere_mesh
 
 
 def test_sphere_plotter_initialisation_one_mesh(dummy_mesh):
@@ -692,6 +710,166 @@ def test_rotate_to_view_theta_radians(mock_nested_histogram_meshes):
     plotter.rotate_to_view(theta=theta_value, use_degrees=use_degrees)
 
     assert np.isclose(plotter.current_theta, np.degrees(theta_value))
+
+
+def test_cell_picking_not_active(mock_orientation_histogram_mesh_counts):
+    """Test that initially cell picking is inactive.
+
+    Test for :property:`.SpherePlotter.face_picking_active` when the
+    plotter is first created.
+    """
+
+    plotter = vr.plotting.SpherePlotter(mock_orientation_histogram_mesh_counts)
+    plotter.produce_plot()
+
+    assert not plotter.cell_picking_active
+    assert len(plotter.picked_cells) == 0
+
+
+def test_cell_picking_active(mock_orientation_histogram_mesh_counts):
+    """Test for activation of cell picking.
+
+    Test for :property:`.SpherePlotter.cell_picking_active` when activating
+    cell picking.
+    """
+
+    plotter = vr.plotting.SpherePlotter(mock_orientation_histogram_mesh_counts)
+    plotter.produce_plot()
+
+    plotter.cell_picking_active = True
+
+    assert plotter.cell_picking_active
+
+
+def test_picked_cells(mock_orientation_histogram_mesh_counts):
+    """Test for picking cells.
+
+    Test for programmatically picking cells through the interactive render
[…]
