### 2025-07-04 1f93ead
test: add test for deselecting cell in `SpherePlotter`.
