### 2025-07-04 04ce117
test: add tests for getting vectors from selected faces.
