### 2025-07-04 53e9c05
docs: add `NaN` clarification for log scale plotting.
