### 2025-07-04 e6f9cd3
fix: fix bug that flipped spatial coordinates when creating symmetric vectors.
