### 2025-07-07 e55b2d1
feat: add public interface to programmatically select cells.
