### 2025-07-07 c002dc4
test: integrate new picking interface into unit tests.
