### 2025-07-07 b85a14d
feat: add method for converting orientation bins into cell mesh indices.
