### 2025-07-07 abd34d9
test: add unit tests for index calculation.
