### 2025-07-07 8037e49
fix: fix bugs related to using single vectors in `util` functions.
