### 2025-07-07 19e1571
docs: replace `property` role with `attr` and fix ambiguous reference.
