### 2025-07-07 0c358f0
docs: add new page to users' guide describing vector filtering.
