+++ docs/users_guide/vector_filtering.md
@@ -1,0 +1,546 @@
+---
+jupytext:
+  text_representation:
+    extension: .md
+    format_name: myst
+    format_version: 0.13
+    jupytext_version: 1.17.1
+kernelspec:
+  display_name: Python 3 (ipykernel)
+  language: python
+  name: python3
+---
+
+# Vector Filtering
+
+```{code-cell} ipython3
+:tags: [remove-cell]
+
+# Configure PyVista behind the scenes before starting
+import matplotlib.pyplot as plt
+import pandas as pd
+import pyvista as pv
+
+try:
+    pv.start_xvfb()
+except OSError:
+    pass
+
+pv.set_jupyter_backend("html")
+pv.global_theme.font.fmt = "%.6g"
+
+pd.options.display.max_rows = 20
+```
+
+At this point, we've now seen how to generate rich histogram visualisations
+that allow us to observe patterns in a collection of vectors. But, what if
+we want to take things a step further and only study vectors with a
+magnitude in a specific range, or with a specific orientation?
+
+In this section, we'll see how to *filter* vector datasets based on
+magnitude and orientation.
+
+```{important}
+When we use the term *filter*, we are referring to selecting specific
+vectors on the basis of magnitude and/or orientation.
+```
+
+We'll continue with our {download}`two_clusters.npy <./two_clusters.npy>`
+dataset, which can also be loaded directly in VectoRose without any
+separate download using {attr}`.SampleData.TWO_CLUSTERS` in the
+{mod}`.data` module.
+
+As usual, let's start by loading the data.
+
+```{code-cell} ipython3
+import numpy as np # We'll use NumPy a bit later
+
+import vectorose as vr
+import vectorose.data
+
+my_vectors = vr.data.SampleData.TWO_CLUSTERS.load()
+my_vectors = vr.util.remove_zero_vectors(my_vectors)
+
+my_vectors
+```
+
+As usual, let's also create a {class}`.FineTregenzaSphere` with 10 nested
+shells to use for bin assignment in terms of magnitude and orientation.
+
+```{code-cell} ipython3
+my_sphere = vr.tregenza_sphere.FineTregenzaSphere(number_of_shells=10)
+labelled_vectors, magnitude_bins = my_sphere.assign_histogram_bins(my_vectors)
+
+labelled_vectors
+```
+
+Our vectors have now been assigned to bins based on magnitude and
+orientation. With this in mind, we can already see a couple of possible
+ways that we may want to filter:
+
+* based on **magnitude** using the shell index, or
+* based on **orientation** using the angular bin index.
+
+With these ideas in mind, let's get started on filtering!
+
+```{tip}
+This page is not a comprehensive guide on filtering. Since vector datasets
+are represented as NumPy arrays and pandas DataFrames, the possibilities
+are quite endless. This page seeks to show some simple workflows for
+filtering using the representations and functions included in VectoRose.
+```
+
+Now that we have the vectors assigned to histogram bins, we can use the
+built-in tools provided by [**pandas**](https://pandas.pydata.org/) to
+perform filtering based on both magnitude and orientation. We'll also see
+some novel tools provided by VectoRose to simplify this filtering process.
+
+Before we get started, here's what the nested histogram shells for this
+dataset look like:
+
+```{video} ./assets/shells_video/shells_video.mp4
+:width: 100%
+:autoplay:
+:loop:
+:poster: ./assets/shells_video/shells_video.png
+:alt: Example video of the histogram shells.
+```
+
+We'll keep this distribution in mind as we perform filtering.
+
+## Magnitude Filtering
+
+The vector magnitude is represented by a scalar value. Filtering based on
[…]
