### 2025-07-07 2d55ccb
1.4.0
Automatically generated by python-semantic-release
