### 2025-07-07 #8369 @nkrusch
> Sorry for the delay. Thanks [@stevenpawley](https://github.com/stevenpawley) for the suggestion. I'll try incorporate a couple of lines about R packages. Given the widespread use of R in the statistics community, I think that would make the discussion more complete. I'll recompile the paper and I'll let you know once I've made the changes.

@bzrudski are these pending changes complete?
