### 2025-07-07 #8369 @bzrudski
> > … Given the widespread use of R in the statistics community, I think that would make the discussion more complete. I'll recompile the paper and I'll let you know once I've made the changes.
>
> [@bzrudski](https://github.com/bzrudski) are these pending changes complete?

@nkrusch Thanks for reaching out. I am trying to finish the by tomorrow. I have also managed to add a couple of new features to the software over the past couple of weeks. Can I mention those in the text, or should I not since they were added during the review, separate from the reviewers comments?
