### 2025-07-07 #8369 @nkrusch
@bzrudski OK - just ping me when the changes are ready and I will do the post-review checks after that.

Regarding the new features: continued development of the software is obviously always good and encouraged. However, the paper should remain consistent with what was already reviewed. I recommend considering a JOSS paper as one point in time, over the evolution of the software, and with awareness that the software is expected to evolve after the paper.
