### 2025-07-08 #8369 @bzrudski
Sounds good. I've sent the changes to my co-authors. I'll let you know once I've pushed the new paper version.

In terms of the software version, I think we can set it to `1.2.0`. This version was created after merging https://github.com/bzrudski/vectorose/pull/8 and closing https://github.com/bzrudski/vectorose/issues/7, based on the reviewer suggestions.
