### 2025-07-08 db55a43
docs: mention R package per reviewer comments.
