+++ paper/paper.md
@@ -43,7 +43,7 @@ To gain such insight, we seek to visualise and quantitatively analyse sets of 3D
 
 ![Scalar data, consisting of real numbers, can be visualised using a 1D scalar histogram, while unit vectors, or *pure directions* and *pure orientations* can be visualised using spherical histograms. **VectoRose** introduces a novel type of histogram to visualise non-unit vectors, consisting of *nested spherical histograms*, that couples these two metrics.\label{fig:stat-support}](./figures/stat-support/stat-support.png)
 
-Other works have previously implemented similar directional analysis approaches. @hacihabibogluSphstatPython32022 developed `sphstat`, a Python package implementing many of the methods described by @fisherStatisticalAnalysisSpherical1993. While `sphstat` provides many statistical approaches, it can only produce 2D scatter plots, not histograms. Three-dimensional histogram plotting is offered in the `spherical_stats` package developed by @schmitzSpherical_stats2021. However, that package only implements a limited subset of quantitative analyses. In addition, the 3D plotting method used relies on a congruent form of sphere patching, which results in discrepancies in face area [@beckersUniversalProjectionComputing2011]. This issue is mitigated in the `sphere-histo` software developed by the @3d-pliSpherehisto2022 group, which constructs spherical histograms using a triangulated sphere representation. Unfortunately, that software only generates histograms and does not offer any ability to integrate this plotting functionality into larger packages. Similarly, while OrientationPy developed by @vasileOrientationPy2022 offers better approaches for constructing 2D spherical plots using equal-area binning, as well as 3D spherical histograms, it also does not include any other statistical analyses. In addition, these packages provide few automated software tests, complicating bug detection. In contrast to these existing tools, VectoRose offers the ability to visualise and quantitatively analyse collections of unit and non-unit vectors in 3D, with intuitive Python functions, most of which are automatically tested.
[…]
