### 2025-07-08 #8369 @nkrusch
@bzrudski do not worry about the date in the `paper.md`. The dates that appear in the rendered version are automatically set as we go through the review process. Currently:

The published date will be populated later.
