### 2025-07-08 #8369 @bzrudski
Here's the updated proof.
