### 2025-07-08 49a5b91
docs: fix case on reference titles.
