+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @software{3d-pliSpherehisto2022,
-  title = {Sphere-Histo},
+  title = {{{sphere-histo}}},
   author = {{3D-PLI}},
   date = {2022-01-31},
   url = {https://github.com/3d-pli/sphere-histo}
@@ -67,7 +67,7 @@
 }
 
 @software{hacihabibogluSphstatPython32022,
-  title = {Sphstat: {{A Python}} 3 Package for Spherical Statistics},
+  title = {{{sphstat}}: {{A Python}} 3 Package for Spherical Statistics},
   author = {Hacihabiboglu, Huseyin},
   date = {2022-12-28},
   url = {https://github.com/hacihabiboglu/sphstat},
@@ -158,7 +158,7 @@
 }
 
 @software{schmitzSpherical_stats2021,
-  title = {Spherical\_stats},
+  title = {{{spherical\_stats}}},
   author = {Schmitz, Daniel},
   date = {2021-11-30},
   url = {https://github.com/dschmitz89/spherical_stats},
