### 2025-07-08 #8369 @bzrudski
Just fixed the case of some of the reference titles.
