### 2025-07-09 #8369 @stevenpawley
> [@nkrusch](https://github.com/nkrusch), I've pushed the changes. [@stevenpawley](https://github.com/stevenpawley), let me know what you think of the revision. Thank you again for your help and suggestions.
>
> Also, [@nkrusch](https://github.com/nkrusch), should I update the date in `paper.md` (it is given as 21 April 2025)?

The revision looks great, congratulations on a nice paper and package.
