### 2025-07-09 a74262d
build: add version placeholder to CHANGELOG.md.
