### 2025-07-09 4b1054a
docs: manually update CHANGELOG.md.
