+++ CHANGELOG.md
@@ -2,6 +2,151 @@
 
 <!-- version list -->
 
+## v1.4.0 (2025-07-07)
+
+_This release is published under the MIT License._
+
+### Bug Fixes
+
+- Change picked cells to have a consistent magenta colour. ([`a6ef9ab`](https://github.com/bzrudski/vectorose/commit/a6ef9ab5fe8ed45f54e97a9b53a8dd7fa2d8a2dd))
+
+- Fix bugs related to using single vectors in `util` functions. ([`8037e49`](https://github.com/bzrudski/vectorose/commit/8037e49669939820caff3c68fbb278e70c1d3b54))
+
+- Rename the scalar fields for the `TregenzaSphere` and `TriangleSphere`. ([`285a90e`](https://github.com/bzrudski/vectorose/commit/285a90e3ddb6f121921672fda57eab109cc12648))
+
+- Switch clicking to right mouse button for simplicity. ([`464fd37`](https://github.com/bzrudski/vectorose/commit/464fd374ec75a64d3b94823bcff5bcfcae570f48))
+
+### Code Style
+
+- Minor code reformatting using `black`. ([`947431d`](https://github.com/bzrudski/vectorose/commit/947431d53a61073df26db80881d1c907c5cf63f7))
+
+### Documentation
+
+- Add new page to users' guide describing vector filtering. ([`0c358f0`](https://github.com/bzrudski/vectorose/commit/0c358f0ab366a712bd4bb2e26edeca5ba89783b3))
+
+- Fix a typo in the main landing page. ([`0abeb3e`](https://github.com/bzrudski/vectorose/commit/0abeb3e47a9082708e339e1b68a6b00ef575ad35))
+
+- Replace `property` role with `attr` and fix ambiguous reference. ([`19e1571`](https://github.com/bzrudski/vectorose/commit/19e1571f921272d719d325c9c6976e828ea1723d))
+
+### Features
+
+- Add method for converting orientation bins into cell mesh indices. ([`b85a14d`](https://github.com/bzrudski/vectorose/commit/b85a14de242b99d63749b874cd02a753fb5ea731))
+
+- Add public interface to programmatically select cells. ([`e55b2d1`](https://github.com/bzrudski/vectorose/commit/e55b2d184d911bc87150b641dfea6cdf7ea7244c))
+
+- Add support for picking mesh cells in `SpherePlotter`. ([`7b04b2a`](https://github.com/bzrudski/vectorose/commit/7b04b2a208bc5603b0243aec5afcf334f910e887))
+
+- Add the ability to filter the vectors from selected histogram cells. ([`1c54f59`](https://github.com/bzrudski/vectorose/commit/1c54f592498a3a05bae615c4168bb2298caa0b6f))
+
+### Refactoring
+
+- Rename functions using cell instead of face. ([`eded369`](https://github.com/bzrudski/vectorose/commit/eded369ec7dc65206ce66780f07e055eca43df6d))
+
+### Testing
+
+- Add test for deselecting cell in `SpherePlotter`. ([`1f93ead`](https://github.com/bzrudski/vectorose/commit/1f93ead6e96e43336af5eab8a453ab20fc061406))
+
+- Add tests for cell picking in the `SpherePlotter`. ([`fa86e67`](https://github.com/bzrudski/vectorose/commit/fa86e67573989d68134151c180150ecab2e9844b))
+
+- Add tests for getting vectors from selected faces. ([`04ce117`](https://github.com/bzrudski/vectorose/commit/04ce1173211a292b9e6bda6d169ebc6be02e8aba))
+
[…]
