### 2025-07-09 a2d99ee
docs: fix line breaks in CHANGELOG.md.
