+++ CHANGELOG.md
@@ -8,51 +8,69 @@ _This release is published under the MIT License._
 
 ### Bug Fixes
 
-- Change picked cells to have a consistent magenta colour. ([`a6ef9ab`](https://github.com/bzrudski/vectorose/commit/a6ef9ab5fe8ed45f54e97a9b53a8dd7fa2d8a2dd))
+- Change picked cells to have a consistent magenta colour.
+  ([`a6ef9ab`](https://github.com/bzrudski/vectorose/commit/a6ef9ab5fe8ed45f54e97a9b53a8dd7fa2d8a2dd))
 
-- Fix bugs related to using single vectors in `util` functions. ([`8037e49`](https://github.com/bzrudski/vectorose/commit/8037e49669939820caff3c68fbb278e70c1d3b54))
+- Fix bugs related to using single vectors in `util` functions.
+  ([`8037e49`](https://github.com/bzrudski/vectorose/commit/8037e49669939820caff3c68fbb278e70c1d3b54))
 
-- Rename the scalar fields for the `TregenzaSphere` and `TriangleSphere`. ([`285a90e`](https://github.com/bzrudski/vectorose/commit/285a90e3ddb6f121921672fda57eab109cc12648))
+- Rename the scalar fields for the `TregenzaSphere` and `TriangleSphere`.
+  ([`285a90e`](https://github.com/bzrudski/vectorose/commit/285a90e3ddb6f121921672fda57eab109cc12648))
 
-- Switch clicking to right mouse button for simplicity. ([`464fd37`](https://github.com/bzrudski/vectorose/commit/464fd374ec75a64d3b94823bcff5bcfcae570f48))
+- Switch clicking to right mouse button for simplicity.
+  ([`464fd37`](https://github.com/bzrudski/vectorose/commit/464fd374ec75a64d3b94823bcff5bcfcae570f48))
 
 ### Code Style
 
-- Minor code reformatting using `black`. ([`947431d`](https://github.com/bzrudski/vectorose/commit/947431d53a61073df26db80881d1c907c5cf63f7))
+- Minor code reformatting using `black`.
+  ([`947431d`](https://github.com/bzrudski/vectorose/commit/947431d53a61073df26db80881d1c907c5cf63f7))
 
 ### Documentation
 
-- Add new page to users' guide describing vector filtering. ([`0c358f0`](https://github.com/bzrudski/vectorose/commit/0c358f0ab366a712bd4bb2e26edeca5ba89783b3))
+- Add new page to users' guide describing vector filtering.
+  ([`0c358f0`](https://github.com/bzrudski/vectorose/commit/0c358f0ab366a712bd4bb2e26edeca5ba89783b3))
 
-- Fix a typo in the main landing page. ([`0abeb3e`](https://github.com/bzrudski/vectorose/commit/0abeb3e47a9082708e339e1b68a6b00ef575ad35))
+- Fix a typo in the main landing page.
+  ([`0abeb3e`](https://github.com/bzrudski/vectorose/commit/0abeb3e47a9082708e339e1b68a6b00ef575ad35))
 
-- Replace `property` role with `attr` and fix ambiguous reference. ([`19e1571`](https://github.com/bzrudski/vectorose/commit/19e1571f921272d719d325c9c6976e828ea1723d))
+- Replace `property` role with `attr` and fix ambiguous reference.
+  ([`19e1571`](https://github.com/bzrudski/vectorose/commit/19e1571f921272d719d325c9c6976e828ea1723d))
 
 ### Features
 
[…]
