### 2025-07-10 #8369 @bzrudski
I'm going back in time on the command line to create the new release. I'm just making one small change to the associated docs (the changelog wasn't updating properly, so I'm manually updating it to add a few release notes).
