### 2025-07-10 #8369 @nkrusch
@bzrudski - I'm currently doing the final editorial checks. The next steps for you will be (this will answer your question):

1. Make a tagged release of your software
2. Archive the reviewed software in Zenodo (or a similar service)
3. List the version tag of the archived version here.
4. Check the archival deposit (in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it), and authors' ORCID.
5. Please list the DOI of the archived version here.
