### 2025-07-10 b46ca9a
docs: update affiliations on paper.
