### 2025-07-10 #8369 @bzrudski
Archive metadata is now up to date.
* DOI for the latest version is: https://doi.org/10.5281/zenodo.15857820
* DOI for `v1.2.0` is: https://doi.org/10.5281/zenodo.15857821
