### 2025-07-16 5753c36
docs: fix reference `lombaertHumanAtlasCardiac2012`.
