+++ docs/refs.bib
@@ -96,17 +96,19 @@
   langid = {english}
 }
 
-@article{lombaertHumanAtlasCardiac2012b,
+@article{lombaertHumanAtlasCardiac2012,
   title = {Human Atlas of the Cardiac Fiber Architecture: {{Study}} on a Healthy Population},
-  author = {Lombaert, Herve and Peyrat, Jean-Marc and Croisille, Pierre and Rapacchi, Stanislas and Fanton, Laurent and Cheriet, Farida and Clarysse, Patrick and Magnin, Isabelle and Delingette, Herve and Ayache, Nicholas},
+  shorttitle = {Human Atlas of the Cardiac Fiber Architecture},
+  author = {Lombaert, Herve and Peyrat, Jean-Marc and Croisille, Pierre and Rapacchi, Stanislas and Fanton, Laurent and Cheriet, Farida and Clarysse, Patrick and Magnin, Isabelle and Delingette, Herv{\'e} and Ayache, Nicholas},
   year = {2012},
+  month = jul,
   journal = {IEEE Transactions on Medical Imaging},
   volume = {31},
   number = {7},
   pages = {1436--1447},
-  issn = {02780062},
-  doi = {10.1109/TMI.2012.2192743},
-  abstract = {Cardiac fibers, as well as their local arrangement in laminar sheets, have a complex spatial variation of their orientation that has an important role in mechanical and electrical cardiac functions. In this paper, a statistical atlas of this cardiac fiber architecture is built for the first time using human datasets. This atlas provides an average description of the human cardiac fiber architecture along with its variability within the population. In this study, the population is composed of ten healthy human hearts whose cardiac fiber architecture is imaged ex vivo with DT-MRI acquisitions. The atlas construction is based on a computational framework that minimizes user interactions and combines most recent advances in image analysis: graph cuts for segmentation, symmetric log-domain diffeomorphic demons for registration, and log-Euclidean metric for diffusion tensor processing and statistical analysis. Results show that the helix angle of the average fiber orientation is highly correlated to the transmural depth and ranges from-41 on the epicardium to +66 on the endocardium. Moreover, we find that the fiber orientation dispersion across the population (13) is lower than for the laminar sheets (31). This study, based on human hearts, extends previous studies on other mammals with concurring conclusions and provides a description of the cardiac fiber architecture more specific to human and better suited for clinical applications. Indeed, this statistical atlas can help to improve the computational models used for radio-frequency ablation, cardiac resynchronization therapy, surgical ventricular restoration, or diagnosis and followups of heart diseases due to fiber architecture anomalies.  2012 IEEE.}
+  publisher = {{Institute of Electrical and Electronics Engineers Inc.}},
+  issn = {0278-0062, 1558-254X},
+  doi = {10.1109/TMI.2012.2192743}
 }
 
 @book{mardiaDirectionalStatistics2000,
