### 2025-07-16 9049d45
fix: rename `include_spatial_locations` to match subclass implementations.
