### 2025-07-18 f630ce9
docs: add citation file.
