### 2025-07-18 3a06716
docs: add citations to JOSS paper.
