+++ docs/refs.bib
@@ -135,6 +135,18 @@
   abstract = {Abstract Trabecular bone anisotropy, describing preferential trabecular co-alignment, is a proxy for its long-term loading history. Trabecular anisotropy varies locally, thus rendering averaged calculations across an entire bone inutile. Here we present a 3D trabecular anisotropy mapping method using vector fields where each vector reflects the extent of local co-alignment of the elementary units of surface. 3D anisotropy maps of hundreds of thousands of vectors were visualized by their magnitude and direction. Similarly, volume fraction was mapped as 3D scalar fields. We constructed anisotropy and volume fraction maps using micro-computed tomography of four presumably nonpathologic human calcanei and compared their anisotropy signature with pathologically loaded calcanei in club foot and calcaneonavicular ankylosis. In the nonpathologic calcaneus, a pattern of four anisotropy trajectories (bands) was consistently identified as dorsal, plantar, Achilles', and peroneal bands. Both pathologic specimens deviated from the nonpathologic maps. The calcaneus in the congenitally disused club foot showed very low local anisotropy values, no co-oriented bands, and low volume fraction. The ankylosed calcaneus showed lower anisotropy than the nonpathologic calcaneus, but not to the same extent as the club foot, and showed patchy high volume fraction. The directionality of co-oriented bands was barely discernable in the ankylosed calcaneus as compared to nonpathologic calcaneus. The anisotropy signature of the nonpathologic calcaneus is consistent with a kinetic loading pattern attributable to walking. The loss of this kinetic loading results in an absent/vanishing anisotropy signature. Such 3D mapping adds new dimensions to quantitative bioimaging of bone and the understanding of skeletal adaptation.}
 }
 
+@article{rudskiVectoRoseNewPackage2025,
+  title = {{{VectoRose}}: {{A}} New Package for Analysing and Visualising {{3D}} Non-Unit Vectors in {{Python}}},
+  author = {Rudski, Benjamin Z. and Deering, Joseph and Reznikov, Natalie},
+  year = {2025},
+  journal = {Journal of Open Source Software},
+  volume = {10},
+  number = {111},
+  pages = {8369},
+  publisher = {The Open Journal},
+  doi = {10.21105/joss.08369}
+}
+
 @article{tregenzaSubdivisionSkyHemisphere1987,
   title = {Subdivision of the Sky Hemisphere for Luminance Measurements},
   author = {Tregenza, P.R.},
