### 2025-07-18 754075a
docs: add DOI badge.
