### 2025-07-24 33920c8
docs: correct Beckers and Beckers reference (and add details to Lombaert et al.).
