+++ paper/paper.bib
@@ -21,19 +21,19 @@
   urldate = {2024-10-02}
 }
 
-@article{beckersUniversalProjectionComputing2011,
-  title = {The Universal Projection for Computing Data Carried on the Hemisphere},
-  author = {Beckers, Benoit and Masset, Luc and Beckers, Pierre},
-  date = {2011-02},
-  journaltitle = {Computer-Aided Design},
-  shortjournal = {Computer-Aided Design},
-  volume = {43},
-  number = {2},
-  pages = {219--226},
-  issn = {00104485},
-  doi = {10.1016/j.cad.2010.11.007},
-  url = {https://linkinghub.elsevier.com/retrieve/pii/S0010448510002174},
-  urldate = {2024-08-16},
+@article{beckersGeneralRuleDisk2012,
+  title = {A General Rule for Disk and Hemisphere Partition into Equal-Area Cells},
+  author = {Beckers, Benoit and Beckers, Pierre},
+  date = {2012-08},
+  journaltitle = {Computational Geometry},
+  shortjournal = {Computational Geometry},
+  volume = {45},
+  number = {7},
+  pages = {275--283},
+  issn = {09257721},
+  doi = {10.1016/j.comgeo.2012.01.011},
+  url = {https://linkinghub.elsevier.com/retrieve/pii/S0925772112000296},
+  urldate = {2024-02-07},
   langid = {english}
 }
 
@@ -109,7 +109,7 @@
 @article{lombaertHumanAtlasCardiac2012,
   title = {Human Atlas of the Cardiac Fiber Architecture: {{Study}} on a Healthy Population},
   shorttitle = {Human Atlas of the Cardiac Fiber Architecture},
-  author = {Lombaert, H. and Peyrat, J. and Croisille, P. and Rapacchi, S. and Fanton, L. and Cheriet, F. and Clarysse, P. and Magnin, I. and Delingette, H. and Ayache, N.},
+  author = {Lombaert, Herve and Peyrat, Jean-Marc and Croisille, Pierre and Rapacchi, Stanislas and Fanton, Laurent and Cheriet, Farida and Clarysse, Patrick and Magnin, Isabelle and Delingette, Herv\'e and Ayache, Nicholas},
   date = {2012-07},
   journaltitle = {IEEE Transactions on Medical Imaging},
   shortjournal = {IEEE Trans. Med. Imaging},
