### 2025-07-24 #8369 @bzrudski
Hello @crvernon. Thank you for all your help with my paper. I just noticed that I put in an incorrect reference at one spot in the paper (most of the authors are the same, and the publication years are very close). I've updated `paper.md` and `paper.bib` in a recent commit (https://github.com/bzrudski/vectorose/commit/33920c87f343766333924670ba5ee033b176a7b7). Is it possible to update the published version of the paper? My sincere apologies for the oversight and any inconvenience.
