### 2025-08-01 041c0e5
build: add PyTables as dependency to allow exporting HDF5 files.
