### 2025-08-01 c2ee240
docs: add access date for Lombaert et al. (2012).
