+++ docs/refs.bib
@@ -108,7 +108,8 @@
   pages = {1436--1447},
   publisher = {{Institute of Electrical and Electronics Engineers Inc.}},
   issn = {0278-0062, 1558-254X},
-  doi = {10.1109/TMI.2012.2192743}
+  doi = {10.1109/TMI.2012.2192743},
+  urldate = {2022-06-15}
 }
 
 @book{mardiaDirectionalStatistics2000,
