### 2025-08-01 18174c7
docs: add warning to mock_data.py about needing to explicity import.
