### 2025-10-23 455d712
test: add unit test of re-picking previously picked cells.
