### 2025-10-23 2e3e4fc
1.4.2
Automatically generated by python-semantic-release
