### 2025-10-28 af93b46
feat: add ability to position camera using Euler angles.
