### 2025-10-28 c858a03
test: add unit tests for the Euler angles (roll not yet tested).
