### 2025-10-28 4e7f2f6
feat: add getters for the azimuth, elevation and roll.
