### 2025-10-28 57f7651
feat: add tests for the azimuth, elevation and roll getters.
