### 2025-10-28 bf98087
1.5.0
Automatically generated by python-semantic-release
