### 2025-10-28 dbef186
fix: update plotter window after rotating camera.
