### 2025-10-28 b2d7cc8
style: reformat code with black.
