### 2025-10-28 0054d88
1.5.1
Automatically generated by python-semantic-release
